+++ CHANGELOG.md
@@ -18,6 +18,14 @@ All notable changes to the "cosmy-wasmy" extension will be documented in this fi
 ### Security 
 -->
 
+## [Unreleased]
+
+### Changed
+
+- Implemented [#2](https://github.com/spoo-bar/cosmy-wasmy/issues/2) - Show faucet icon for account only when faucet has been set up for that chain in the settings
+
+
+
 ## [v0.2.0] - 15 June 2022
 
 ### Added
@@ -33,11 +41,15 @@ All notable changes to the "cosmy-wasmy" extension will be documented in this fi
 
 - Fixed [#1](https://github.com/spoo-bar/cosmy-wasmy/issues/1) - Using contract address as id (instead of codeId) for the rendered item in the Contract view
 
+
+
 ## [v0.1.1] - 03 June 2022
 
 ### Changed
 
 - Changed the extension settings. Split the gas price value into seperate price and denom
+
+
 
 ## [v0.1.0] - 02 June 2022